부산대 BE_박혜연 5주차 과제 (1단계) (#137)
* feat: 4주차 멘토님 피드백 반영

* docs: README.md

* chores: .gitignore

* feat: 카카오 로그인 인가코드 받기

* feat: 카카오 로그인(토큰발급)

* feat: RestClientConfig

* test: kakaoTests

* refactor: 멘토님 피드백 반영

* chores: jsonNaming 추가

+++ src/main/java/gift/administrator/option/OptionApiController.java
@@ -1,12 +1,14 @@
 package gift.administrator.option;
 
+import jakarta.validation.Valid;
 import java.util.List;
-import org.springframework.data.crossstore.ChangeSetPersister.NotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -32,10 +34,20 @@ public class OptionApiController {
     }
 
     @DeleteMapping("/options/{optionId}")
-    public ResponseEntity<OptionDTO> deleteOptionByOptionId(
-        @PathVariable("optionId") Long optionId)
-        throws NotFoundException {
+    public ResponseEntity<Void> deleteOptionByOptionId(
+        @PathVariable("optionId") long optionId) {
+        if (optionService.countAllOptionsByProductIdFromOptionId(optionId) == 1) {
+            throw new IllegalArgumentException("상품에는 적어도 하나의 옵션이 있어야 합니다.");
+        }
         optionService.deleteOptionByOptionId(optionId);
         return ResponseEntity.status(HttpStatus.OK).build();
+    }
+
+    @PutMapping("/products/{productId}/options/{optionId}")
+    public ResponseEntity<OptionDTO> updateOptionByProductIdAndOptionId(
+        @PathVariable("productId") long productId, @PathVariable("optionId") long optionId,
+        @Valid @RequestBody OptionDTO optionDTO) {
+        optionService.existsByNameSameProductIdNotOptionId(optionDTO.getName(), productId, optionId);
+        return ResponseEntity.ok(optionService.updateOption(optionId, optionDTO));
     }
 }
